better structure the code
