+++ src/components/FigSource/index.vue
@@ -1,0 +1,16 @@
+<template src="./template.html"></template>
+<script src="./script.js"></script>
+
+<style scoped>
+	.el-tabs {
+		height: 40vh;
+		font-family: Helvetica;
+		border-style: groove 1px;
+		margin: 5px;
+		border-radius: 8px;
+	}
+
+	.el-tab-pane {
+		border-radius: 8px;
+	}
+</style>
