+++ src/store/state.js
@@ -1,0 +1,40 @@
+const state = {
+    items: [{
+        content: 'Get Milk two',
+        attachedEle: [],
+        positions: [],
+        selected: false,
+    },
+    {
+        content: 'Get Water one',
+        attachedEle: [],
+        positions: [],
+        selected: false,
+    },
+    {
+        content: 'Get Bread three',
+        attachedEle: [],
+        positions: [],
+        selected: false,
+    },
+    {
+        content: 'Get shampoo four',
+        attachedEle: [],
+        positions: [],
+        selected: false,
+    },
+    {
+        content: 'Get juice three',
+        attachedEle: [],
+        positions: [],
+        selected: false,
+    },
+    ],
+    newItem: {
+        content: 'be new new new',
+        attachedEle: [],
+        selected: false,
+    },
+    selectedItem: {},
+};
+export default state;
